ai/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.0
greenlet==3.2.3
h11==0.16.0
holidays==0.77
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mlxtend==0.23.4
mpmath==1.3.0
narwhals==2.0.1
numpy==2.2.6
onnxruntime==1.22.1
openai==1.98.0
packaging==25.0
pandas==2.3.1
pdfkit==1.0.0
pillow==11.3.0
plotly==6.2.0
prophet==1.1.7
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
stanio==0.5.1
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
